Share directory component for browser and finder
Change the browser and the finder to share the directory component
