+++ src/components/browse/mod.rs
@@ -1,12 +1,8 @@
 mod browse;
 mod contents;
-mod directory;
 
 pub use browse::{Browse, Effect as BrowseEffect, Event as BrowseEvent, Props as BrowseProps};
 use contents::{
     Contents as ContentsComponent, Effect as ContentsEffect, Event as ContentsEvent,
     Props as ContentsProps,
-};
-use directory::{
-    Directory as DirectoryComponent, Effect as DirectoryEffect, Event as DirectoryEvent,
 };
