+++ src/components/finder/mod.rs
@@ -1,9 +1,7 @@
-mod directory;
 mod finder;
 mod found;
 mod phrase;
 
-use directory::{Directory, Props as DirectoryProps};
 pub use finder::{Effect as FinderEffect, Finder, Props as FinderProps};
 use found::{Effect as FoundEffect, Event as FoundEvent, Found, Props as FoundProps};
 use phrase::{Effect as PhraseEffect, Event as PhraseEvent, Phrase};
